Add implementation of exchange related requests and improve the company ones
- Now controls concurrency accessing company resources info,
because dropwizard uses a thread pool to handle requests and
we don't gave a database for experimental purposes.
- Now implements all exchange related requests on the designed API, except one
related to companies present on an exchange. This is because there isn't enough
information to process this effeciently for now.
- Also controls concurrency accessing exchange related resources for the same reason
 referred on the first topic.
- Removed ExchangeId class because it made no sense to have one class just for
that purpose. Now uses ObjectMapper and ObjectNode from jackson to create JSON
with the exchange id.